+++ mne/viz/tests/test_epochs.py
@@ -11,7 +11,7 @@ import warnings
 from nose.tools import assert_raises
 
 import numpy as np
-
+from numpy.testing import assert_equal
 
 from mne import io, read_events, Epochs
 from mne import pick_types
@@ -131,6 +131,18 @@ def test_plot_epochs_image():
     import matplotlib.pyplot as plt
     epochs = _get_epochs()
     plot_epochs_image(epochs, picks=[1, 2])
+    overlay_times = [0.1]
+    plot_epochs_image(epochs, order=[0], overlay_times=overlay_times)
+    plot_epochs_image(epochs, overlay_times=overlay_times)
+    assert_raises(ValueError, plot_epochs_image, epochs,
+                  overlay_times=[0.1, 0.2])
+    assert_raises(ValueError, plot_epochs_image, epochs,
+                  order=[0, 1])
+    with warnings.catch_warnings(record=True) as w:
+        plot_epochs_image(epochs, overlay_times=[1.1])
+        warnings.simplefilter('always')
+    assert_equal(len(w), 1)
+
     plt.close('all')