+++ doc/whats_new.rst
@@ -10,7 +10,7 @@ Current
 Changelog
 ~~~~~~~~~
 
-    -
+    - Add `overlay_times` parameter to :func:`mne.viz.plot_epochs_image` to be able to display for example reaction times on top of the images, by `Alex Gramfort`_
 
 BUG
 ~~~